refactored routing
