+++ package.json
@@ -20,11 +20,13 @@
     "react": "^18.2.0",
     "react-appointment-picker": "^3.2.1",
     "react-dom": "^18.2.0",
-    "react-error-boundary": "^3.1.4"
+    "react-error-boundary": "^3.1.4",
+    "react-router-dom": "^6.8.1"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
+    "@types/react-router-dom": "^5.3.3",
     "@types/amplify": "^1.1.25",
     "@types/jest": "^29.4.0",
     "@types/jsdom": "^21.1.0",
