+++ src/components/ui-components/PrescriptionUpdateForm.d.ts
@@ -7,35 +7,38 @@
 import * as React from "react";
 import { GridProps, TextFieldProps } from "@aws-amplify/ui-react";
 import { EscapeHatchProps } from "@aws-amplify/ui-react/internal";
-import { Todo } from "../models";
+import { Prescription } from "../../util/models";
 export declare type ValidationResponse = {
     hasError: boolean;
     errorMessage?: string;
 };
 export declare type ValidationFunction<T> = (value: T, validationResponse: ValidationResponse) => ValidationResponse | Promise<ValidationResponse>;
-export declare type TodoUpdateFormInputValues = {
+export declare type PrescriptionUpdateFormInputValues = {
     name?: string;
+    dosage?: string;
     description?: string;
 };
-export declare type TodoUpdateFormValidationValues = {
+export declare type PrescriptionUpdateFormValidationValues = {
     name?: ValidationFunction<string>;
+    dosage?: ValidationFunction<string>;
     description?: ValidationFunction<string>;
 };
 export declare type PrimitiveOverrideProps<T> = Partial<T> & React.DOMAttributes<HTMLDivElement>;
-export declare type TodoUpdateFormOverridesProps = {
-    TodoUpdateFormGrid?: PrimitiveOverrideProps<GridProps>;
+export declare type PrescriptionUpdateFormOverridesProps = {
+    PrescriptionUpdateFormGrid?: PrimitiveOverrideProps<GridProps>;
     name?: PrimitiveOverrideProps<TextFieldProps>;
+    dosage?: PrimitiveOverrideProps<TextFieldProps>;
     description?: PrimitiveOverrideProps<TextFieldProps>;
 } & EscapeHatchProps;
-export declare type TodoUpdateFormProps = React.PropsWithChildren<{
-    overrides?: TodoUpdateFormOverridesProps | undefined | null;
+export declare type PrescriptionUpdateFormProps = React.PropsWithChildren<{
+    overrides?: PrescriptionUpdateFormOverridesProps | undefined | null;
 } & {
     id?: string;
-    todo?: Todo;
-    onSubmit?: (fields: TodoUpdateFormInputValues) => TodoUpdateFormInputValues;
-    onSuccess?: (fields: TodoUpdateFormInputValues) => void;
-    onError?: (fields: TodoUpdateFormInputValues, errorMessage: string) => void;
-    onChange?: (fields: TodoUpdateFormInputValues) => TodoUpdateFormInputValues;
-    onValidate?: TodoUpdateFormValidationValues;
+    prescription?: Prescription;
+    onSubmit?: (fields: PrescriptionUpdateFormInputValues) => PrescriptionUpdateFormInputValues;
+    onSuccess?: (fields: PrescriptionUpdateFormInputValues) => void;
+    onError?: (fields: PrescriptionUpdateFormInputValues, errorMessage: string) => void;
+    onChange?: (fields: PrescriptionUpdateFormInputValues) => PrescriptionUpdateFormInputValues;
+    onValidate?: PrescriptionUpdateFormValidationValues;
 } & React.CSSProperties>;
-export default function TodoUpdateForm(props: TodoUpdateFormProps): React.ReactElement;
+export default function PrescriptionUpdateForm(props: PrescriptionUpdateFormProps): React.ReactElement;
