+++ src/pages/Pharmacist.tsx
@@ -1,8 +1,15 @@
 import { Button, useAuthenticator, View } from '@aws-amplify/ui-react';
 import React, { useRef, useState } from 'react';
 import * as AwsUI from '@awsui/components-react';
+import { Navigation } from '../components/navigation/Navigation';
+import type { AmplifyUser, AuthEventData } from '@aws-amplify/ui';
 
-function Pharmacist({ user }: { user: any }) {
+export interface PharmacistProps {
+  user?: AmplifyUser;
+  signOut?: (event?: AuthEventData) => void;
+}
+
+function Pharmacist({ user }: PharmacistProps) {
   const { signOut } = useAuthenticator((context) => [context.user]);
   const [navigationOpen, setNavigationOpen] = React.useState(false);
   const [isLoading, setLoading] = useState(false);
@@ -12,19 +19,10 @@ function Pharmacist({ user }: { user: any }) {
     <AwsUI.AppLayout
       navigation={
         <>
-          <View className="Account">
-            Account: {user.attributes.email}
-            <br />
-            Role: {user.attributes['custom:role']}
-          </View>
-
-          <Button className="Signout" onClick={signOut}>
-            {' '}
-            Sign out{' '}
-          </Button>
+          <Navigation signOut={signOut} />
         </>
       }
-      content={<></>}
+      content={<>Pharmacist</>}
     />
   );
 }
